requirements.txt
APScheduler==3.9.0.post1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==7.1.2
cryptography==36.0.1
Flask==1.1.2
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.28
oauthlib==3.2.0
paramiko==2.9.2
protobuf==3.19.4
pycparser==2.21
PyNaCl==1.5.0
pyperclip==1.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sshtunnel==0.4.0
tweepy==4.6.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==1.0.1
